+++ src/professions/admin/dto/legislation.dto.ts
@@ -1,4 +1,4 @@
-import { IsNotEmpty } from 'class-validator';
+import { IsNotEmpty, IsUrl, ValidateIf } from 'class-validator';
 
 export default class LegislationDto {
   @IsNotEmpty({
@@ -6,10 +6,29 @@ export default class LegislationDto {
   })
   nationalLegislation: string;
 
-  @IsNotEmpty({
-    message: 'professions.form.errors.legislation.link.empty',
-  })
+  @IsUrl(
+    {},
+    {
+      message: 'professions.form.errors.legislation.link.invalid',
+    },
+  )
   link: string;
 
-  change: boolean;
+  @IsNotEmpty({
+    message:
+      'professions.form.errors.legislation.secondNationalLegislation.empty',
+  })
+  @ValidateIf((e) => e.secondLink || e.secondNationalLegislation)
+  secondNationalLegislation?: string;
+
+  @IsUrl(
+    {},
+    {
+      message: 'professions.form.errors.legislation.secondLink.invalid',
+    },
+  )
+  @ValidateIf((e) => e.secondLink || e.secondNationalLegislation)
+  secondLink?: string;
+
+  change: string;
 }