+++ src/professions/admin/interfaces/legislation.template.ts
@@ -2,6 +2,7 @@ import { Legislation } from '../../../legislations/legislation.entity';
 
 export interface LegislationTemplate {
   legislation: Legislation | null;
+  secondLegislation: Legislation | null;
   captionText: string;
   change: boolean;
   errors?: object;