+++ src/professions/admin/legislation.controller.spec.ts
@@ -36,31 +36,69 @@ describe(LegislationController, () => {
   });
 
   describe('edit', () => {
-    it('renders the Legislation page, passing in any values on the Profession that have already been set', async () => {
-      const legislation = legislationFactory.build({
-        name: 'Legal Services Act 2007',
-        url: 'www.example.com',
+    describe('when the Profession contains a single legislation', () => {
+      it('renders the Legislation page, passing in any values on the Profession that have already been set', async () => {
+        const legislation = legislationFactory.build({
+          name: 'Legal Services Act 2007',
+          url: 'www.example.com',
+        });
+        const profession = professionFactory.build({
+          id: 'profession-id',
+        });
+
+        const version = professionVersionFactory.build({
+          id: 'version-id',
+          legislations: [legislation],
+        });
+
+        professionsService.findWithVersions.mockResolvedValue(profession);
+        professionVersionsService.findWithProfession.mockResolvedValue(version);
+
+        await controller.edit(response, 'profession-id', 'version-id', 'false');
+
+        expect(response.render).toHaveBeenCalledWith(
+          'admin/professions/legislation',
+          expect.objectContaining({
+            legislation: legislation,
+          }),
+        );
       });
-      const profession = professionFactory.build({
-        id: 'profession-id',
+    });
+
+    describe('when the Profession contains a second Legislation', () => {
+      it('renders the Legislation page, passing in any values on the Profession that have already been set', async () => {
+        const legislation1 = legislationFactory.build({
+          name: 'Legal Services Act 2007',
+          url: 'www.example.com',
+        });
+
+        const legislation2 = legislationFactory.build({
+          name: 'Another Legal Services Act 2007',
+          url: 'www.another-example.com',
+        });
+
+        const profession = professionFactory.build({
+          id: 'profession-id',
+        });
+
+        const version = professionVersionFactory.build({
+          id: 'version-id',
+          legislations: [legislation1, legislation2],
+        });
+
+        professionsService.findWithVersions.mockResolvedValue(profession);
+        professionVersionsService.findWithProfession.mockResolvedValue(version);
+
+        await controller.edit(response, 'profession-id', 'version-id', 'false');
+
+        expect(response.render).toHaveBeenCalledWith(
+          'admin/professions/legislation',
+          expect.objectContaining({
+            legislation: legislation1,
+            secondLegislation: legislation2,
+          }),
+        );
       });
-
-      const version = professionVersionFactory.build({
-        id: 'version-id',
-        legislations: [legislation],
-      });
-
-      professionsService.findWithVersions.mockResolvedValue(profession);
-      professionVersionsService.findWithProfession.mockResolvedValue(version);
-
-      await controller.edit(response, 'profession-id', 'version-id', false);
-
-      expect(response.render).toHaveBeenCalledWith(
-        'admin/professions/legislation',
-        expect.objectContaining({
-          legislation: legislation,
-        }),
-      );
     });
   });
 
@@ -76,7 +114,7 @@ describe(LegislationController, () => {
         const dto: LegislationDto = {
           link: 'www.legal-legislation.com',
           nationalLegislation: 'Legal Services Act 2008',
-          change: false,
+          change: 'false',
         };
 
         professionsService.findWithVersions.mockResolvedValue(profession);
@@ -102,6 +140,49 @@ describe(LegislationController, () => {
       });
     });
 
+    describe('when a second Legislation is entered', () => {
+      it('creates two new Legislations on the Profession and redirects to Check your answers', async () => {
+        const profession = professionFactory.build({ id: 'profession-id' });
+        const version = professionVersionFactory.build({
+          id: 'version-id',
+          profession: profession,
+        });
+
+        const dto: LegislationDto = {
+          link: 'www.legal-legislation.com',
+          nationalLegislation: 'Legal Services Act 2008',
+          secondLink: 'www.another-legal-legislation.com',
+          secondNationalLegislation: 'Another Legal Services Act 2008',
+          change: 'false',
+        };
+
+        professionsService.findWithVersions.mockResolvedValue(profession);
+        professionVersionsService.findWithProfession.mockResolvedValue(version);
+
+        await controller.update(response, 'profession-id', 'version-id', dto);
+
+        expect(professionVersionsService.save).toHaveBeenCalledWith(
+          expect.objectContaining({
+            legislations: [
+              expect.objectContaining({
+                url: 'www.legal-legislation.com',
+                name: 'Legal Services Act 2008',
+              }),
+              expect.objectContaining({
+                url: 'www.another-legal-legislation.com',
+                name: 'Another Legal Services Act 2008',
+              }),
+            ],
+            profession: profession,
+          }),
+        );
+
+        expect(response.redirect).toHaveBeenCalledWith(
+          '/admin/professions/profession-id/versions/version-id/check-your-answers',
+        );
+      });
+    });
+
     describe('when required parameters are not entered', () => {
       it('renders the edit page with errors and does not update the Profession', async () => {
         const profession = professionFactory.build({ id: 'profession-id' });
@@ -113,7 +194,7 @@ describe(LegislationController, () => {
         const dto: LegislationDto = {
           link: undefined,
           nationalLegislation: undefined,
-          change: false,
+          change: 'false',
         };
 
         professionsService.findWithVersions.mockResolvedValue(profession);
@@ -128,7 +209,7 @@ describe(LegislationController, () => {
           expect.objectContaining({
             errors: {
               link: {
-                text: 'professions.form.errors.legislation.link.empty',
+                text: 'professions.form.errors.legislation.link.invalid',
               },
               nationalLegislation: {
                 text: 'professions.form.errors.legislation.nationalLegislation.empty',